feat: new button for analyzing survey responses
